requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asyncer==0.0.7
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2024.8.30
chainlit==1.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.112.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
jiter==0.5.0
joblib==1.4.2
langfuse==2.50.2
Lazify==0.4.0
literalai==0.0.607
llama-cloud==0.0.17
llama-index==0.11.12
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.12
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.9
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.6
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.47.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==23.2
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.37.2
striprtf==0.0.26
syncer==2.0.3
tenacity==8.5.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.12.0
zipp==3.20.2
